Организация кода. Модуляция.
