+++ js/data.js
@@ -1,0 +1,11 @@
+
+const DESCRIPTIONS = [
+  'Не послушавшись рукописи, наш текст продолжил свой путь.',
+  'Ценность яблок кроется в их составе',
+  'Сложно сказать, почему небо темнеет',
+  'Все чаще появляется информация',
+  'Постоянный количественный рост станет частью наших традиций',
+  'Внезапно, солнечных дней всё меньше'
+
+];
+export {DESCRIPTIONS};
